[Opt](func) opt the percentile func performance
