+++ be/src/vec/functions/simple_function_factory.h
@@ -108,8 +108,8 @@ class SimpleFunctionFactory {
     using Creator = std::function<FunctionBuilderPtr()>;
     using FunctionCreators = phmap::flat_hash_map<std::string, Creator>;
     using FunctionIsVariadic = phmap::flat_hash_set<std::string>;
-    /// @TEMPORARY: for be_exec_version=4
-    constexpr static int NEWEST_VERSION_FUNCTION_SUBSTITUTE = 4;
+    /// @TEMPORARY: for be_exec_version=5
+    constexpr static int NEWEST_VERSION_FUNCTION_SUBSTITUTE = 5;
 
 public:
     void register_function(const std::string& name, const Creator& ptr) {
@@ -148,7 +148,7 @@ public:
     /// @TEMPORARY: for be_exec_version=3
     template <class Function>
     void register_alternative_function() {
-        static std::string suffix {"_old_for_version_before_4_0"};
+        static std::string suffix {"_old_for_version_before_5_0"};
         function_to_replace[Function::name] = Function::name + suffix;
         register_function(Function::name + suffix, &createDefaultFunction<Function>);
     }
@@ -192,7 +192,7 @@ private:
     FunctionCreators function_creators;
     FunctionIsVariadic function_variadic_set;
     std::unordered_map<std::string, std::string> function_alias;
-    /// @TEMPORARY: for be_exec_version=3. replace function to old version.
+    /// @TEMPORARY: for be_exec_version=4. replace function to old version.
     std::unordered_map<std::string, std::string> function_to_replace;
 
     template <typename Function>
@@ -200,7 +200,7 @@ private:
         return std::make_shared<DefaultFunctionBuilder>(Function::create());
     }
 
-    /// @TEMPORARY: for be_exec_version=3
+    /// @TEMPORARY: for be_exec_version=4
     void temporary_function_update(int fe_version_now, std::string& name) {
         // replace if fe is old version.
         if (fe_version_now < NEWEST_VERSION_FUNCTION_SUBSTITUTE &&
